fixed legal
